backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.11
pandas==2.3.2
passlib==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
